+++ contracts/protocol/libraries/helpers/Quantity.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.0;
 library Quantity {
     function getPartiallyLiquidate(
         int256 _quantity,
-        int256 _liquidationPenaltyRatio
+        uint256 _liquidationPenaltyRatio
     ) internal pure returns (int256) {
-        return (_quantity * _liquidationPenaltyRatio) / 100;
+        return (_quantity * int256(_liquidationPenaltyRatio)) / 100;
     }
 
     function isSameSide(int256 qA, int256 qB) internal pure returns (bool) {
